requirements.txt
Flask==3.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
python-telegram-bot==20.7
python-dotenv==1.0.0

